destroy project route that destroys associated cards and join table calls
